Use coarser processing method for Data processing inequality test
because rounding and flooring may increase the entropy via numeric
artifacts. This is less likely with using sign and should make the
test less flaky.

Except, of course, there was an error i did not see, in which case
this change is bad as the test in total becomes more robust.