(fix) Remove float from bin coercion
Single float is not supported to specify bins. It could be unclear
if it is a bin edge or the number of bins. People should either
add an integer of the number of bins, a list of integers (number
of bins) or one or many arrays (that can be of floats) to give the
bin edges in all or per dimension.